Restructure tttc-common for dependabot expectations
