+++ common/firebase/index.ts
