+++ common/morphisms/tree-metrics.ts
@@ -1,4 +1,4 @@
-import * as schema from "../schema";
+import * as schema from "../schema/";
 import { z } from "zod";
 
 // ! these functions haven't been tested thoroughly yet
